requirements.txt
mp-api
pymatgen
matminer
pandas
python-dotenv
scikit-learn
matplotlib
